chore: freeze the current state for pre-phase 1
In preparation for the new direction of the site build, all in progress
files have been updated to minimum viable states and frozen. GitHub will
no longer be the source of truth in the new direction of this project's
architecture.

See `docs/project-scope.md` or the `README.md` if details are not found
in the `project-scope.md`.